add form add note and separate FAB
