+++ src/screens/FormNote.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import {
+  View,
+  Text
+} from 'react-native';
+
+class FormNote extends React.Component{
+
+    render(){
+        return(
+            <View>
+                
+                <Text>Test</Text>
+
+            </View>
+        )
+    }
+
+}
+
+export default FormNote;
