+++ src/screens/MainPage.js
@@ -7,7 +7,7 @@ import {
 
 import NoteItem from '../components/item/note_item'
 import EmptyNote from '../components/empty/empty_note'
-import Icon from 'react-native-vector-icons/FontAwesome';
+import FloatingActionButton from '../components/button/floating_action_button'
   
 class MainPage extends React.Component{
     
@@ -45,7 +45,7 @@ class MainPage extends React.Component{
                         bottom: 10,
                         right: 10,
                     }}
-                    onPress={() => alert(`I'm being clicked!`)}
+                    onPress={() => this.props.navigation.navigate('Form')}
                 />
             </View>
         );
@@ -53,20 +53,3 @@ class MainPage extends React.Component{
 }
 
 export default MainPage;
-
-const FloatingActionButton = (props) => (
-    <TouchableOpacity onPress={props.onPress} style={props.style}>
-      <View
-        style={{
-          backgroundColor: 'purple',
-          width: 45,
-          height: 45,
-          borderRadius: 45,
-          justifyContent:'center',
-          alignItems:'center'
-        }}
-      >
-           <Icon name="plus" size={16} color="white" />
-        </View>
-    </TouchableOpacity>
-  );
